feat: 여러개의 서버를 사용할 수 있는 기능 구현
