+++ src/pages/ErrorPage.tsx
@@ -1,4 +1,4 @@
-import { useNavigate } from 'react-router-dom';
+import { useNavigate, useRouteError } from 'react-router-dom';
 import { styled } from 'styled-components';
 import Icon from '../assets/images/icon.png';
 
@@ -9,7 +9,9 @@ const Content = styled.main`
   justify-content: center;
   gap: 24px;
 
-  margin-top: 60px;
+  margin: 60px auto auto 0;
+
+  max-width: 600px;
 `;
 
 const Cart = styled.img`
@@ -29,12 +31,15 @@ const Button = styled.button`
 
 const ErrorPage = () => {
   const navigate = useNavigate();
+  const error = useRouteError();
 
   return (
     <Content>
       <Cart src={Icon} alt="아이콘" />
 
-      <Title>찾으시는 페이지가 없는 것 같아요! 🥲</Title>
+      <Title>페이지를 표시하는 중 오류가 발생했거나, 찾으시는 페이지가 없는 것 같아요! 🥲</Title>
+
+      <pre>{String(error)}</pre>
 
       <Button onClick={() => navigate('/')}>홈으로 가기</Button>
     </Content>
