+++ src/recoil/atoms/productsState.ts
@@ -1,13 +1,21 @@
-import { atom } from 'recoil';
-import { client } from '../../api';
+import { atomFamily, selector } from 'recoil';
+import type Client from '../../api/Client';
+import type { ShoppingCartRestAPI } from '../../api/rest/ShoppingCartRestAPI';
 import type { Product } from '../../type';
+import productsQuery from '../queries/productsQuery';
+import clientState from './clientState';
 
-const productsState = atom<Product[]>({
+const localProductsState = atomFamily<Product[], Client<ShoppingCartRestAPI>>({
+  key: 'localProductsState',
+  default: productsQuery,
+});
+
+const productsState = selector<Product[]>({
   key: 'productsState',
-  default: client
-    .get('/products')
-    .acceptOrThrow(200)
-    .then((response) => response.data),
+  get: ({ get }) => {
+    const client = get(clientState);
+    return get(localProductsState(client));
+  },
 });
 
 export default productsState;
